Unbiased Random Samples and Permutations

Unbiased Random Samples and Permutations

dqsample(x, size, replace = FALSE, prob = NULL)

dqsample.int(n, size = n, replace = FALSE, prob = NULL)

Arguments

x
either a vector of one or more elements from which to choose, or a positive integer.

size
a non-negative integer giving the number of items to choose.

replace
should sampling be with replacement?

prob
a vector of probability weights for obtaining the elements of the vector being sampled.

n
a positive number, the number of items to choose from.

See also