Generate seed as a integer vector

Generate seed as a integer vector

generateSeedVectors(nseeds, nwords = 2L)

Arguments

nseeds

Integer scalar, number of seeds to generate.

nwords

Integer scalar, number of words to generate per seed.

Value

A list of length n, where each element is an integer vector that contains nwords words (i.e., 32*nwords bits) of randomness.

Details

Each seed is encoded as an integer vector with the most significant bits at the start of the vector. Each integer vector is converted into an unsigned integer (in C++ or otherwise) by the following procedure:

Start with a sum of zero.

Add the first value of the vector.

Left-shift the sum by 32.

Add the next value of the vector, and repeat.

The aim is to facilitate R-level generation of seeds with sufficient randomness to cover the entire state space of pseudo-random number generators that require more than the ~32 bits available in an int. It also preserves the integer nature of the seed, thus avoiding problems with casting double-precision numbers to integers.

It is possible for the seed vector to contain NA_integer_ values. This should not be cause for alarm, as R uses -INT_MAX to encode missing values in integer vectors.

Author

Aaron Lun

Examples

generateSeedVectors(10, 2)
#> [[1]]
#> [1] 873887222 782093289
#> 
#> [[2]]
#> [1]  318412058 1293975181
#> 
#> [[3]]
#> [1]  -335320052 -1774518192
#> 
#> [[4]]
#> [1] -1961913921  -575952814
#> 
#> [[5]]
#> [1]  451211419 1847755042
#> 
#> [[6]]
#> [1] -479365647 1176069025
#> 
#> [[7]]
#> [1] -1357510172  1844824771
#> 
#> [[8]]
#> [1] 1813231165 1163883152
#> 
#> [[9]]
#> [1] -209713366 -340609233
#> 
#> [[10]]
#> [1] 319407724 968473996
#> 

generateSeedVectors(5, 4)
#> [[1]]
#> [1]  1396926767 -1594425359 -1258752060 -2114556933
#> 
#> [[2]]
#> [1]  782431055 1355383180  103525712  250672588
#> 
#> [[3]]
#> [1]   827737070    37404664 -1581670144  1938918286
#> 
#> [[4]]
#> [1]  122543331 -363746787 2008604274  985660363
#> 
#> [[5]]
#> [1]  -934887333 -1230426903  -359009901  1050797445
#>