Generate seed as a integer vector

Generate seed as a integer vector

generateSeedVectors(nseeds, nwords = 2L)

Arguments

nseeds
Integer scalar, number of seeds to generate.

nwords
Integer scalar, number of words to generate per seed.

Value

A list of length n, where each element is an integer vector that contains nwords words (i.e., 32*nwords bits) of randomness.

Details

Each seed is encoded as an integer vector with the most significant bits at the start of the vector. Each integer vector is converted into an unsigned integer (in C++ or otherwise) by the following procedure:

Start with a sum of zero.

Add the first value of the vector.

Left-shift the sum by 32.

Add the next value of the vector, and repeat.

The aim is to facilitate R-level generation of seeds with sufficient randomness to cover the entire state space of pseudo-random number generators that require more than the ~32 bits available in an int. It also preserves the integer nature of the seed, thus avoiding problems with casting double-precision numbers to integers.

It is possible for the seed vector to contain NA_integer_ values. This should not be cause for alarm, as R uses -INT_MAX to encode missing values in integer vectors.

Examples

generateSeedVectors(10, 2)

#> [[1]]
#> [1]   346819306 -1714718956
#> 
#> [[2]]
#> [1]   31780411 2137937573
#> 
#> [[3]]
#> [1] -1147263134  -538586090
#> 
#> [[4]]
#> [1]  147065385 1727986585
#> 
#> [[5]]
#> [1] 1733182984 1669459452
#> 
#> [[6]]
#> [1]  1245077974 -1136793703
#> 
#> [[7]]
#> [1] -83581474 220960067
#> 
#> [[8]]
#> [1] -1306426408   134133148
#> 
#> [[9]]
#> [1] 1292058449 2057394743
#> 
#> [[10]]
#> [1] -1260856967   774549019
#> 

generateSeedVectors(5, 4)

#> [[1]]
#> [1] -1373689793 -1538975404   412420705  -989239170
#> 
#> [[2]]
#> [1] -126611791 1980780712  750227268 2120154851
#> 
#> [[3]]
#> [1]   876939355   280101037  -750762845 -1846642902
#> 
#> [[4]]
#> [1] -1734038120  -224352187 -1955913565  1918572463
#> 
#> [[5]]
#> [1]  120521019 1675171970 1619077831 -613821391
#>