Generate seed as a integer vector

Generate seed as a integer vector

generateSeedVectors(nseeds, nwords = 2L)

Arguments

nseeds

Integer scalar, number of seeds to generate.

nwords

Integer scalar, number of words to generate per seed.

Value

A list of length n, where each element is an integer vector that contains nwords words (i.e., 32*nwords bits) of randomness.

Details

Each seed is encoded as an integer vector with the most significant bits at the start of the vector. Each integer vector is converted into an unsigned integer (in C++ or otherwise) by the following procedure:

Start with a sum of zero.

Add the first value of the vector.

Left-shift the sum by 32.

Add the next value of the vector, and repeat.

The aim is to facilitate R-level generation of seeds with sufficient randomness to cover the entire state space of pseudo-random number generators that require more than the ~32 bits available in an int. It also preserves the integer nature of the seed, thus avoiding problems with casting double-precision numbers to integers.

It is possible for the seed vector to contain NA_integer_ values. This should not be cause for alarm, as R uses -INT_MAX to encode missing values in integer vectors.

Author

Aaron Lun

Examples

generateSeedVectors(10, 2)
#> [[1]]
#> [1] -1929658750 -1234047105
#> 
#> [[2]]
#> [1]  -797777734 -2137835429
#> 
#> [[3]]
#> [1] 1756396954  -36317834
#> 
#> [[4]]
#> [1]   981766944 -1771878444
#> 
#> [[5]]
#> [1]  -112613599 -1717841157
#> 
#> [[6]]
#> [1] -1010395690  -252272857
#> 
#> [[7]]
#> [1]  1878728873 -1123466506
#> 
#> [[8]]
#> [1] 469262666  50376032
#> 
#> [[9]]
#> [1] 1695217884 -706485578
#> 
#> [[10]]
#> [1] -1634573493  1040827914
#> 

generateSeedVectors(5, 4)
#> [[1]]
#> [1] 1238789451  -48591202 1262649099  336382876
#> 
#> [[2]]
#> [1]  1165938687 -1311532745   387115523  -836044773
#> 
#> [[3]]
#> [1] -1524164930   329514465   859330424  -451804651
#> 
#> [[4]]
#> [1]   957216103    21296662  -325435167 -1070053661
#> 
#> [[5]]
#> [1]   700323073 -1032812034  1595785105 -2142751478
#>